move TagsContext to hooks
